@@ -1,115 +1,81 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { RootState, selectors } from '../store';
-import { IonPopover, IonIcon, IonSelect, IonSelectOption, IonHeader, IonToolbar, IonButtons, IonButton, IonMenuButton, IonContent, IonList, IonItem, IonLabel, IonDatetime, IonTitle } from '@ionic/react';
-import './About.css';
+import React, { useState } from 'react';
+import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonButton, IonIcon, IonDatetime, IonSelectOption, IonList, IonItem, IonLabel, IonSelect, IonPopover } from '@ionic/react';
+import './About.scss';
+import { calendar, pin, more } from 'ionicons/icons';
 import AboutPopover from '../components/AboutPopover';
-import { calendar, pin } from 'ionicons/icons';
 
-type Props = ReturnType<typeof mapStateToProps>
+interface AboutProps { }
 
-type State = {
-  showPopover: boolean,
-  showPopoverEvent: null | MouseEvent
-}
+const About: React.FC<AboutProps> = () => {
 
-class About extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
+  const [showPopover, setShowPopover] = useState(false);
+  const [popoverEvent, setPopoverEvent] = useState();
 
-    this.state = {
-      showPopover: false,
-      showPopoverEvent: null
-    };
-  }
+  const presentPopover = (e: React.MouseEvent) => {
+    setPopoverEvent(e.nativeEvent);
+    setShowPopover(true);
+  };
+  const conferenceDate = '2047-05-17';
 
-  presentPopover = (e: MouseEvent) => {
-    this.setState(() => ({
-      showPopover: true,
-      showPopoverEvent: e
-    }));
-  }
+  return (
+    <IonPage id="about-page">
+      <IonHeader>
+        <IonToolbar>
+          <IonButtons slot="start">
+            <IonMenuButton></IonMenuButton>
+          </IonButtons>
+          <IonTitle>About</IonTitle>
+          <IonButtons slot="end">
+            <IonButton icon-only onClick={presentPopover}>
+              <IonIcon slot="icon-only" icon={more}></IonIcon>
+            </IonButton>
+          </IonButtons>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent>
 
-  dismissPopover = () => {
-    this.setState(() => ({
-      'showPopover': false,
-      'showPopoverEvent': null
-    }));
-  }
+        <div className="about-header">
+          <img src="assets/img/ionic-logo-white.svg" alt="ionic logo" />
+        </div>
+        <div className="about-info">
+          <h4 className="ion-padding-start">Ionic Conference</h4>
 
-  render() {
-    return (
-      <>
-        <IonHeader>
-          <IonToolbar color="primary">
-            <IonButtons slot="start">
-              <IonMenuButton></IonMenuButton>
-            </IonButtons>
-            <IonTitle>About</IonTitle>
-            <IonButtons slot="end">
-              <IonButton icon-only onClick={this.presentPopover}>
-                <IonIcon slot="icon-only" name="more"></IonIcon>
-              </IonButton>
-            </IonButtons>
-          </IonToolbar>
-        </IonHeader>
+          <IonList lines="none">
+            <IonItem>
+              <IonIcon icon={calendar} slot="start"></IonIcon>
+              <IonLabel position="stacked">Date</IonLabel>
+              <IonDatetime displayFormat="MMM DD, YYYY" max="2056" value={conferenceDate}></IonDatetime>
+            </IonItem>
 
-        <IonPopover
-          isOpen={this.state.showPopover}
-          event={this.state.showPopoverEvent}
-          onDidDismiss={this.dismissPopover}
-        >
-          <AboutPopover
-            dismissPopover={this.dismissPopover}
-          />
-        </IonPopover>
+            <IonItem>
+              <IonIcon icon={pin} slot="start"></IonIcon>
+              <IonLabel position="stacked">Location</IonLabel>
+              <IonSelect>
+                <IonSelectOption value="madison" selected>Madison, WI</IonSelectOption>
+                <IonSelectOption value="austin">Austin, TX</IonSelectOption>
+                <IonSelectOption value="chicago">Chicago, IL</IonSelectOption>
+                <IonSelectOption value="seattle">Seattle, WA</IonSelectOption>
+              </IonSelect>
+            </IonItem>
+          </IonList>
 
-        <IonContent>
-          <div className="about-header">
-            <img src="assets/img/ionic-logo-white.svg" alt="ionic logo" />
-          </div>
-          <div className="ion-padding about-info">
-            <h4>Ionic Conference</h4>
+          <p className="ion-padding-start ion-padding-end">
+            The Ionic Conference is a one-day conference featuring talks from the Ionic team. It is focused on Ionic applications being
+            built with Ionic 2. This includes migrating apps from Ionic 1 to Ionic 2, Angular concepts, Webpack, Sass, and many
+            other technologies used in Ionic 2. Tickets are completely sold out, and we’re expecting more than 1000 developers
+            – making this the largest Ionic conference ever!
+          </p>
+        </div>
+      </IonContent>
+      <IonPopover
+        isOpen={showPopover}
+        event={popoverEvent}
+        onDidDismiss={() => setShowPopover(false)}
+      >
+        <AboutPopover dismiss={() => setShowPopover(false)} /> 
+      </IonPopover>
+    </IonPage>
+  );
+};
 
-            <IonList lines="none">
-              <IonItem>
-                <IonIcon icon={calendar} slot="start"></IonIcon>
-                <IonLabel>Date</IonLabel>
-                <IonDatetime displayFormat="MMM DD, YYYY" max="2056" value={this.props.conferenceDate}></IonDatetime>
-              </IonItem>
-
-              <IonItem>
-                <IonIcon icon={pin} slot="start"></IonIcon>
-                <IonLabel>Location</IonLabel>
-                <IonSelect>
-                  <IonSelectOption value="madison" selected>Madison, WI</IonSelectOption>
-                  <IonSelectOption value="austin">Austin, TX</IonSelectOption>
-                  <IonSelectOption value="chicago">Chicago, IL</IonSelectOption>
-                  <IonSelectOption value="seattle">Seattle, WA</IonSelectOption>
-                </IonSelect>
-              </IonItem>
-            </IonList>
-
-            <p>
-              The Ionic Conference is a one-day conference featuring talks from the
-              Ionic team. It is focused on Ionic applications being built with
-              Ionic 2. This includes migrating apps from Ionic 1 to Ionic 2,
-              Angular concepts, Webpack, Sass, and many other technologies used
-              in Ionic 2. Tickets are completely sold out, and we’re expecting
-              more than 1000 developers – making this the largest Ionic
-              conference ever!
-            </p>
-          </div>
-        </IonContent>
-      </>
-    );
-  }
-}
-
-const mapStateToProps = (state: RootState) => ({
-  conferenceDate: selectors.sessions.conferenceStart(state.sessions),
-});
-
-export default connect(
-  mapStateToProps
-)(About);
+export default About;
