@@ -1,73 +1,108 @@
-import React, { Component } from 'react';
-import { IonHeader, IonToolbar, IonButtons, IonButton, IonTitle, IonContent, IonList, IonItem, IonLabel, IonInput, IonRow, IonCol, IonMenuButton } from '@ionic/react';
-import './Login.css';
+import React, { useState } from 'react';
+import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonRow, IonCol, IonButton, IonList, IonItem, IonLabel, IonInput, IonText } from '@ionic/react';
+import './Login.scss';
+import { setIsLoggedIn, setUsername } from '../data/actions';
+import { connect } from '../data/connect';
+import { RouteComponentProps } from 'react-router';
 
-type State = {
-  username: string | null
+interface OwnProps extends RouteComponentProps {}
+
+interface DispatchProps {
+  setIsLoggedIn: typeof setIsLoggedIn;
+  setUsername: typeof setUsername;
 }
 
-export default class Login extends Component<{}, State> {
-  constructor(props: {}) {
-    super(props);
-    this.state = {
-      username: null
+interface LoginProps extends OwnProps,  DispatchProps { }
+
+const Login: React.FC<LoginProps> = ({setIsLoggedIn, history, setUsername: setUsernameAction}) => {
+
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [formSubmitted, setFormSubmitted] = useState(false);
+  const [usernameError, setUsernameError] = useState(false);
+  const [passwordError, setPasswordError] = useState(false);
+
+  const login = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setFormSubmitted(true);
+    if(!username) {
+      setUsernameError(true);
     }
-  }
+    if(!password) {
+      setPasswordError(true);
+    }
 
-  updateUserName(e: CustomEvent) {}
-  logInUser() {}
-  signUpUser() {}
+    if(username && password) {
+      await setIsLoggedIn(true);
+      await setUsernameAction(username);
+      history.push('/tabs/schedule', {direction: 'none'});
+    }
+  };
 
-  render() {
-    return (
-      <>
-        <IonHeader>
-          <IonToolbar>
-            <IonButtons slot="start">
-              <IonMenuButton></IonMenuButton>
-            </IonButtons>
-            <IonTitle>Login</IonTitle>
-          </IonToolbar>
-        </IonHeader>
+  return (
+    <IonPage id="login-page">
+      <IonHeader>
+        <IonToolbar>
+          <IonButtons slot="start">
+            <IonMenuButton></IonMenuButton>
+          </IonButtons>
+          <IonTitle>Login</IonTitle>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent>
 
-        <IonContent>
-          <div className="logo">
-            <img src="assets/img/appicon.svg" alt="Ionic logo"/>
-          </div>
-          <form noValidate>
-            <IonList no-lines>
-              <IonItem>
-                <IonLabel color="primary">Username</IonLabel>
-                <IonInput
-                  onIonChange={this.updateUserName}
-                  name="username"
-                  type="text"
-                  autocapitalize="off"
-                  value={this.state.username}
-                  required>
-                </IonInput>
-              </IonItem>
-              <IonItem>
-                <IonLabel color="primary">Password</IonLabel>
-                <IonInput name="password" type="password" required></IonInput>
-              </IonItem>
-            </IonList>
+        <div className="login-logo">
+          <img src="assets/img/appicon.svg" alt="Ionic logo" />
+        </div>
 
-            <IonRow responsive-sm>
-              <IonCol>
-                <IonButton onClick={this.logInUser} type="submit">
-                  Login
-                </IonButton>
-              </IonCol>
-              <IonCol>
-                <IonButton onClick={this.signUpUser} color="light">
-                  Signup
-                </IonButton>
-              </IonCol>
-            </IonRow>
-          </form>
-        </IonContent>
-      </>
-    );
-  }
-}
+        <form noValidate onSubmit={login}>
+          <IonList>
+            <IonItem>
+              <IonLabel position="stacked" color="primary">Username</IonLabel>
+              <IonInput name="username" type="text" value={username} spellCheck={false} autocapitalize="off" onIonChange={e => setUsername(e.detail.value!)}
+                required>
+              </IonInput>
+            </IonItem>
+
+            {formSubmitted && usernameError && <IonText color="danger">
+              <p className="ion-padding-start">
+                Username is required
+              </p>
+            </IonText>}
+
+            <IonItem>
+              <IonLabel position="stacked" color="primary">Password</IonLabel>
+              <IonInput name="password" type="password" value={password} onIonChange={e => setPassword(e.detail.value!)}>
+              </IonInput>
+            </IonItem>
+
+            {formSubmitted && passwordError && <IonText color="danger">
+              <p className="ion-padding-start">
+                Password is required
+              </p>
+            </IonText>}
+          </IonList>
+
+          <IonRow>
+            <IonCol>
+              <IonButton type="submit" expand="block">Login</IonButton>
+            </IonCol>
+            <IonCol>
+              <IonButton routerLink="/signup" color="light" expand="block">Signup</IonButton>
+            </IonCol>
+          </IonRow>
+        </form>
+
+      </IonContent>
+
+    </IonPage>
+  );
+};
+
+export default connect<OwnProps, {}, DispatchProps>({
+  mapDispatchToProps: {
+    setIsLoggedIn,
+    setUsername
+  },
+  component: Login
+})
