@@ -1,142 +1,122 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import { RootState, actions } from '../store';
-import { withRouter, RouteComponentProps } from 'react-router';
-import { format } from 'date-fns';
-import { IonLabel, IonItemSliding, IonItem, IonItemOptions, IonItemOption, IonAlert } from '@ionic/react';
-import { Session } from '../store/sessions/types';
-import { AlertButton } from '@ionic/react';
+import React, { useRef, useState } from 'react';
+import { IonItemSliding, IonAlert, IonItem, IonLabel, IonItemOptions, IonItemOption, AlertButton } from '@ionic/react';
+import { Session } from '../models/Session';
+import { format, parseISO as parseDate } from 'date-fns';
+import { connect } from '../data/connect';
+import { addFavorite, removeFavorite } from '../data/actions';
+import { Time } from './Time';
 
-type Props = RouteComponentProps<{}> & typeof mapDispatchToProps & ReturnType<typeof mapStateToProps> & {
+interface OwnProps {
   session: Session;
   listType: "all" | "favorites";
 }
 
-type State = {
-  showAlert: boolean;
-  alertHeader?: string;
-  alertMessage?: string;
-  alertButtons: (AlertButton | string)[];
+interface StateProps {
+  favoriteSessions: number[];
 }
 
-class SessionListItem extends React.Component<Props, State> {
-  ionItemSlidingRef: React.RefObject<any>
-  defaultState: State = {
-    showAlert: false,
-    alertHeader: '',
-    alertMessage: undefined,
-    alertButtons: []
+interface DispatchProps {
+  addFavorite: typeof addFavorite;
+  removeFavorite: typeof removeFavorite;
+}
+
+type Props = OwnProps & StateProps & DispatchProps
+
+const SessionListItemInner: React.FC<Props> = ({ addFavorite, removeFavorite, favoriteSessions, session, listType }) => {
+  const ionItemSlidingRef = useRef<HTMLIonItemSlidingElement>(null)
+  const [showAlert, setShowAlert] = useState(false);
+  const [alertHeader, setAlertHeader] = useState('');
+  const [alertButtons, setAlertButtons] = useState<(AlertButton | string)[]>([]);
+  const isFavorite = favoriteSessions.indexOf(session.id) > -1;
+
+  const dismissAlert = () => {
+    setShowAlert(false);
+    setAlertHeader('');
+    setAlertButtons([]);
+    ionItemSlidingRef.current!.close();
   }
 
-  constructor(props: Props) {
-    super(props);
+  const removeFavoriteSession = () => {
+    addFavorite(session.id);
 
-    this.state = {
-      ...this.defaultState
-    };
-    this.ionItemSlidingRef = React.createRef();
+    // create an alert instance
+    setShowAlert(true);
+    setAlertHeader('Favorite already added');
+    setAlertButtons([
+      {
+        text: 'Cancel',
+        handler: dismissAlert
+      },
+      {
+        text: 'Remove',
+        handler: removeFavorite(session.id)
+      }
+    ]);
   }
 
-  dismissAlert = () => {
-    this.setState(() => ({
-      ...this.defaultState
-    }));
-    this.ionItemSlidingRef.current.close();
-  }
-
-  addFavoriteSession = () => {
-    if (this.props.favoriteSessions.indexOf(this.props.session.id) !== - 1) {
+  const addFavoriteSession = () => {
+    if (isFavorite) {
       // woops, they already favorited it! What shall we do!?
       // prompt them to remove it
-      this.removeFavoriteSession('Favorite already added')();
+      removeFavoriteSession();
     } else {
       // remember this session as a user favorite
-      this.props.addFavorite(this.props.session.id);
+      addFavorite(session.id);
 
       // create an alert instance
-      this.setState({
-        showAlert: true,
-        alertHeader: 'Favorite Added',
-        alertButtons: [
-          {
-            text: 'OK',
-            handler: this.dismissAlert
-          }
-        ]
-      });
-    }
-  }
-
-  removeFavoriteSession = (title: string) => () => {
-    this.setState({
-      showAlert: true,
-      alertHeader: title,
-      alertMessage: 'Would you like to remove this session from your favorites?',
-      alertButtons: [
+      setShowAlert(true);
+      setAlertHeader('Favorite Added');
+      setAlertButtons([
         {
-          text: 'Cancel',
-          handler: this.dismissAlert
-        },
-        {
-          text: 'Remove',
-          handler: () => {
-            this.props.removeFavorite(this.props.session.id);
-            this.dismissAlert();
-          }
+          text: 'OK',
+          handler: dismissAlert
         }
-      ]
-    });
-  }
+      ]);
+    }
+  };
 
-  navigateToSession = (sessionId: number) => () => {
-    this.props.history.push(`/schedule/sessions/${sessionId}`);
-  }
+  return (
+    <IonItemSliding ref={ionItemSlidingRef} class={'track-' + session.tracks[0].toLowerCase()}>
+      <IonAlert
+        isOpen={showAlert}
+        header={alertHeader}
+        buttons={alertButtons}
+        onDidDismiss={dismissAlert}
+      ></IonAlert>
+      <IonItem routerLink={`/tabs/schedule/${session.id}`}>
+        <IonLabel>
+          <h3>{session.name}</h3>
+          <p>
+            <Time date={session.dateTimeStart} /> &mdash;&nbsp;
+            <Time date={session.dateTimeEnd} /> &mdash;&nbsp;
+            {session.location}
+          </p>
+        </IonLabel>
+      </IonItem>
+      <IonItemOptions>
+        {listType === "favorites" ?
+          <IonItemOption color="danger" onClick={() => removeFavoriteSession()}>
+            Remove
+          </IonItemOption>
+          :
+          <IonItemOption color="favorite" onClick={addFavoriteSession}>
+            Favorite
+          </IonItemOption>
+        }
+      </IonItemOptions>
+    </IonItemSliding>
+  );
+};
 
-  render() {
-    return (
-      <IonItemSliding ref={this.ionItemSlidingRef} class={'track-' + this.props.session.tracks[0].toLowerCase()}>
-        <IonAlert
-          isOpen={this.state.showAlert}
-          header={this.state.alertHeader}
-          buttons={this.state.alertButtons}
-          onDidDismiss={this.dismissAlert}
-        ></IonAlert>
-        <IonItem button onClick={this.navigateToSession(this.props.session.id)}>
-          <IonLabel>
-            <h3>{this.props.session.name}</h3>
-            <p>
-              {format(this.props.session.dateTimeStart, "h:MM a")} &mdash;&nbsp;
-              {format(this.props.session.dateTimeEnd, "h:MM a")}:&nbsp;
-              {this.props.session.location}
-            </p>
-          </IonLabel>
-        </IonItem>
-        <IonItemOptions>
-          { this.props.listType === "favorites" ?
-            <IonItemOption color="danger" onClick={this.removeFavoriteSession('Remove Favorite')}>
-              Remove
-            </IonItemOption>
-            :
-            <IonItemOption color="favorite" onClick={this.addFavoriteSession}>
-              Favorite
-            </IonItemOption>
-          }
-        </IonItemOptions>
-      </IonItemSliding>
-    );
-  }
-}
-
-const mapStateToProps = (state: RootState) => ({
-  favoriteSessions: state.sessions.favoriteSessions
+export default connect<OwnProps, StateProps, DispatchProps>({
+  mapStateToProps: (state) => {
+    return {
+      favoriteSessions: state.favorites
+    }
+  },
+  mapDispatchToProps: {
+    addFavorite,
+    removeFavorite
+  },
+  component: SessionListItemInner
 });
-
-const mapDispatchToProps = {
-  addFavorite: (sessionId: number) => actions.sessions.addFavorite(sessionId),
-  removeFavorite: (sessionId: number) => actions.sessions.removeFavorite(sessionId),
-}
-
-export default withRouter(connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(SessionListItem));
