@@ -1,59 +1,108 @@
-import React, { Component } from 'react';
-import { IonHeader, IonToolbar, IonButtons, IonMenuButton, IonTitle, IonContent, IonList, IonItem, IonLabel, IonInput, IonButton } from '@ionic/react';
+import React, { useState } from 'react';
+import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonRow, IonCol, IonButton, IonList, IonItem, IonLabel, IonInput, IonText } from '@ionic/react';
+import './Login.scss';
+import { setIsLoggedIn } from '../data/actions';
+import { connect } from '../data/connect';
+import { RouteComponentProps } from 'react-router';
 
-type State = {
-  username: string | null,
-  password: string | null
+interface OwnProps extends RouteComponentProps {}
+
+interface DispatchProps {
+  setIsLoggedIn: typeof setIsLoggedIn;
 }
 
-export default class Signup extends Component<{},State> {
-  signupFormRef: React.Ref<HTMLFormElement>
+interface LoginProps extends OwnProps,  DispatchProps { }
 
-  constructor(props: {}) {
-    super(props);
-    this.state = {
-      username: null,
-      password: null
+const Login: React.FC<LoginProps> = ({setIsLoggedIn, history}) => {
+
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [formSubmitted, setFormSubmitted] = useState(false);
+  const [usernameError, setUsernameError] = useState(false);
+  const [passwordError, setPasswordError] = useState(false);
+
+  const login = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setFormSubmitted(true);
+    if(!username) {
+      setUsernameError(true);
     }
-    this.signupFormRef = React.createRef();
-  }
+    if(!password) {
+      setPasswordError(true);
+    }
 
-  onSignup() {}
-  render() {
-    return (
-      <>
-        <IonHeader>
-          <IonToolbar>
-            <IonButtons slot="start">
-              <IonMenuButton></IonMenuButton>
-            </IonButtons>
-            <IonTitle>Signup</IonTitle>
-          </IonToolbar>
-        </IonHeader>,
+    if(username && password) {
+      await setIsLoggedIn(true);
+      history.push('/tabs/schedule', {direction: 'none'});
+    }
+  };
 
-        <IonContent class="page-user">
-          <div className="logo">
-            <img src="/assets/img/appicon.svg" alt="Ionic Logo"/>
-          </div>
-          <form ref={this.signupFormRef}>
-            <IonList no-lines>
-              <IonItem>
-                <IonLabel color="primary">Username</IonLabel>
-                <IonInput value={this.state.username} name="username" type="text" required>
-                </IonInput>
-              </IonItem>
-              <IonItem>
-                <IonLabel color="primary">Password</IonLabel>
-                <IonInput value={this.state.password} name="password" type="password" required>
-                </IonInput>
-              </IonItem>
-            </IonList>
-            <div>
-              <IonButton onClick={() => this.onSignup()} type="submit">Create</IonButton>
-            </div>
-          </form>
-        </IonContent>
-      </>
-    );
-  }
-}
+  return (
+    <IonPage id="signup-page">
+      <IonHeader>
+        <IonToolbar>
+          <IonButtons slot="start">
+            <IonMenuButton></IonMenuButton>
+          </IonButtons>
+          <IonTitle>Signup</IonTitle>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent>
+
+        <div className="login-logo">
+          <img src="assets/img/appicon.svg" alt="Ionic logo" />
+        </div>
+
+        <form noValidate onSubmit={login}>
+          <IonList>
+            <IonItem>
+              <IonLabel position="stacked" color="primary">Username</IonLabel>
+              <IonInput name="username" type="text" value={username} spellCheck={false} autocapitalize="off" onIonChange={e => {
+                setUsername(e.detail.value!);
+                setUsernameError(false);
+              }}
+                required>
+              </IonInput>
+            </IonItem>
+
+            {formSubmitted && usernameError && <IonText color="danger">
+              <p className="ion-padding-start">
+                Username is required
+              </p>
+            </IonText>}
+
+            <IonItem>
+              <IonLabel position="stacked" color="primary">Password</IonLabel>
+              <IonInput name="password" type="password" value={password} onIonChange={e => {
+                setPassword(e.detail.value!);
+                setPasswordError(false);
+              }}>
+              </IonInput>
+            </IonItem>
+
+            {formSubmitted && passwordError && <IonText color="danger">
+              <p className="ion-padding-start">
+                Password is required
+              </p>
+            </IonText>}
+          </IonList>
+
+          <IonRow>
+            <IonCol>
+              <IonButton type="submit" expand="block">Create</IonButton>
+            </IonCol>
+          </IonRow>
+        </form>
+
+      </IonContent>
+
+    </IonPage>
+  );
+};
+
+export default connect<OwnProps, {}, DispatchProps>({
+  mapDispatchToProps: {
+    setIsLoggedIn
+  },
+  component: Login
+})
