@@ -1,57 +1,64 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router';
-import { RootState } from '../store';
-import { IonIcon, IonHeader, IonToolbar, IonButtons, IonTitle, IonContent, IonButton, IonBackButton } from '@ionic/react'
-import './SpeakerDetail.css';
+import { IonIcon, IonHeader, IonToolbar, IonButtons, IonTitle, IonContent, IonButton, IonBackButton, IonPage } from '@ionic/react'
+import './SpeakerDetail.scss';
 import { logoTwitter, logoGithub, logoInstagram } from 'ionicons/icons';
+import { connect } from '../data/connect';
+import * as selectors from '../data/selectors';
+import { Speaker } from '../models/Speaker';
 
-type Props = RouteComponentProps<{ id: string, tab: string}> & ReturnType<typeof mapStateToProps> & {
-  goBack: () => void;
+interface OwnProps extends RouteComponentProps {
+  speaker?: Speaker;
 };
 
-const SpeakerDetail: React.SFC<Props> = ({ speakers, match, goBack }) => {
-  const speaker = speakers.find(s => s.id === parseInt(match.params.id, 10));
+interface StateProps {};
+
+interface DispatchProps {};
+
+interface SpeakerDetailProps extends OwnProps, StateProps, DispatchProps {};
+
+const SpeakerDetail: React.FC<SpeakerDetailProps> = ({ speaker }) => {
+  
   if (!speaker) {
-    return null;
+    return <div>Speaker not found</div>
   }
 
   return (
-    <>
+    <IonPage id="speaker-detail">
       <IonHeader>
         <IonToolbar>
           <IonButtons slot="start">
-            <IonBackButton defaultHref={`/${match.params.tab}`} />
+            <IonBackButton defaultHref="/tabs/speakers" />
           </IonButtons>
           <IonTitle>{speaker.name}</IonTitle>
         </IonToolbar>
       </IonHeader>
 
       <IonContent className="ion-padding speaker-detail speaker-page-list">
-        <div>
-          <img src={speaker.profilePic} alt={speaker.name}/>
-          <br/>
-          <IonButton icon-only color="twitter">
-            <IonIcon icon={logoTwitter}></IonIcon>
+        <div className="ion-text-center">
+          <img src={speaker.profilePic} alt={speaker.name} />
+          <br />
+          <IonButton fill="clear" size="small" color="twitter">
+            <IonIcon icon={logoTwitter} slot="icon-only"></IonIcon>
           </IonButton>
-          <IonButton icon-only color="github">
-            <IonIcon icon={logoGithub}></IonIcon>
+          <IonButton fill="clear" size="small" color="github">
+            <IonIcon icon={logoGithub} slot="icon-only"></IonIcon>
           </IonButton>
-          <IonButton icon-only color="instagram">
-            <IonIcon icon={logoInstagram}></IonIcon>
+          <IonButton fill="clear" size="small" color="instagram">
+            <IonIcon icon={logoInstagram} slot="icon-only"></IonIcon>
           </IonButton>
         </div>
 
         <p>{speaker.about}</p>
       </IonContent>
-    </>
+    </IonPage>
   );
 };
 
-const mapStateToProps = (state: RootState) => ({
-  speakers: state.speakers.speakers
-});
 
-export default connect(
-  mapStateToProps
-)(SpeakerDetail)
+export default connect({
+  mapStateToProps: (state, ownProps) => ({
+    speaker: selectors.getSpeaker(state, ownProps)
+  }),
+  component: SpeakerDetail
+});
