@@ -1,37 +1,85 @@
-import React, { Component } from 'react';
-import { IonMenuButton, IonHeader, IonToolbar, IonButtons, IonTitle, IonContent, IonList, IonItem, IonLabel, IonTextarea, IonButton } from '@ionic/react';
+import React, { useState } from 'react';
+import { IonHeader, IonToolbar, IonTitle, IonContent, IonPage, IonButtons, IonMenuButton, IonRow, IonCol, IonButton, IonList, IonItem, IonLabel, IonInput, IonText, IonTextarea, IonToast } from '@ionic/react';
+import './Login.scss';
+import { setIsLoggedIn } from '../data/actions';
+import { connect } from '../data/connect';
+import { RouteComponentProps } from 'react-router';
 
-export default class Support extends Component {
-  submit() {}
-  render() {
-    return (
-      <>
-        <IonHeader>
-          <IonToolbar>
-            <IonButtons slot="start">
-              <IonMenuButton></IonMenuButton>
-            </IonButtons>
-            <IonTitle>Support</IonTitle>
-          </IonToolbar>
-        </IonHeader>
+interface OwnProps { }
 
-        <IonContent class="page-user">
-          <div className="logo">
-            <img src="/assets/img/appicon.svg" alt="Ionic Logo"/>
-          </div>
-          <form onSubmit={() => this.submit()}>
-            <IonList no-lines>
-              <IonItem>
-                <IonLabel color="primary">Enter your support message below</IonLabel>
-                <IonTextarea name="supportQuestion" required></IonTextarea>
-              </IonItem>
-            </IonList>
-            <div>
-              <IonButton type="submit">Submit</IonButton>
-            </div>
-          </form>
-        </IonContent>
-      </>
-    );
-  }
-}
+interface DispatchProps { }
+
+interface SupportProps extends OwnProps, DispatchProps { }
+
+const Support: React.FC<SupportProps> = () => {
+
+  const [message, setMessage] = useState('');
+  const [formSubmitted, setFormSubmitted] = useState(false);
+  const [messageError, setMessageError] = useState(false);
+  const [showToast, setShowToast] = useState(false);
+
+  const send = (e: React.FormEvent) => {
+    e.preventDefault();
+    setFormSubmitted(true);
+    if (!message) {
+      setMessageError(true);
+    }
+    if (message) {
+      setMessage('');
+      setShowToast(true);
+    }
+  };
+
+  return (
+    <IonPage id="support-page">
+      <IonHeader>
+        <IonToolbar>
+          <IonButtons slot="start">
+            <IonMenuButton></IonMenuButton>
+          </IonButtons>
+          <IonTitle>Support</IonTitle>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent>
+
+        <div className="login-logo">
+          <img src="assets/img/appicon.svg" alt="Ionic logo" />
+        </div>
+
+        <form noValidate onSubmit={send}>
+          <IonList>
+            <IonItem>
+              <IonLabel position="stacked" color="primary">Enter your support message below</IonLabel>
+              <IonTextarea name="message" value={message} spellCheck={false} autocapitalize="off" rows={6} onIonChange={e => setMessage(e.detail.value!)}
+                required>
+              </IonTextarea>
+            </IonItem>
+
+            {formSubmitted && messageError && <IonText color="danger">
+              <p className="ion-padding-start">
+                Support message is required
+              </p>
+            </IonText>}
+          </IonList>
+
+          <IonRow>
+            <IonCol>
+              <IonButton type="submit" expand="block">Submit</IonButton>
+            </IonCol>
+          </IonRow>
+        </form>
+       
+      </IonContent>
+     
+      <IonToast
+        isOpen={showToast}
+        duration={3000}
+        message="Your support request has been sent"
+        onDidDismiss={() => setShowToast(false)} />
+    </IonPage>
+  );
+};
+
+export default connect<OwnProps, {}, DispatchProps>({
+  component: Support
+})
