@@ -1,108 +1,87 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { actions } from '../store';
-import { IonIcon, IonHeader, IonToolbar, IonButtons, IonButton, IonContent, IonSlides, IonSlide, IonPage } from '@ionic/react';
-import { withRouter, RouteComponentProps } from 'react-router-dom';
-import './Tutorial.css';
+import React, { useState, useRef } from 'react';
+import { IonContent, IonPage, IonHeader, IonToolbar, IonButtons, IonButton, IonSlides, IonSlide, IonIcon } from '@ionic/react';
+import { arrowForward } from 'ionicons/icons';
+import { setHasSeenTutorial } from '../data/actions';
+import './Tutorial.scss';
+import { connect } from '../data/connect';
+import { RouteComponentProps } from 'react-router';
 
-type Props = RouteComponentProps<{}> & typeof mapDispatchToProps;
+interface OwnProps extends RouteComponentProps {};
 
-type State = {
-  showSkip: boolean
+interface DispatchProps {
+  setHasSeenTutorial: typeof setHasSeenTutorial
 }
 
-class Tutorial extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      showSkip: false
-    }
-  }
+interface TutorialProps extends OwnProps, DispatchProps { };
 
-  onSlideChangeStart = () => {
-    this.setState((state, props) => (
-      {
-        ...state,
-        showSkip: !state.showSkip
-      }
-    ));
-  }
+const Tutorial: React.FC<TutorialProps> = ({ history, setHasSeenTutorial }) => {
+  const [showSkip, setShowSkip] = useState(true);
+  const slideRef = useRef<HTMLIonSlidesElement>(null);
+  
+  const startApp = async () => { 
+    await setHasSeenTutorial(true);
+    history.push('/', { direction: 'none' });
+  };
 
-  endTutorial = () => {
-    this.props.sawTutorial();
-  }
+  const handleSlideChangeStart = () => { 
+    slideRef.current!.isEnd().then(isEnd => setShowSkip(!isEnd));
+  };
 
-  render() {
-    return (
-      <IonPage className="tutorial-page">
-        <IonHeader no-border>
-          <IonToolbar>
-            { this.state.showSkip ?
-              <IonButtons slot="end">
-                <IonButton onClick={this.endTutorial} color="primary">Skip</IonButton>
-              </IonButtons>
-            : null}
-          </IonToolbar>
-        </IonHeader>
-        <IonContent no-bounce>
-          <IonSlides onIonSlideWillChange={this.onSlideChangeStart} pager={false}>
+  return (
+    <IonPage id="tutorial-page">
+      <IonHeader no-border>
+        <IonToolbar>
+          <IonButtons slot="end">
+            {showSkip && <IonButton color='primary' onClick={startApp}>Skip</IonButton>}
+          </IonButtons>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent fullscreen>
 
-            <IonSlide>
-              <img alt="welcome" src="/assets/img/ica-slidebox-img-1.png" className="slide-image"/>
-              <h2 className="slide-title">
-                Welcome to <b>ICA</b>
-              </h2>
-              <p>
-                The <b>ionic conference app</b> is a practical preview of the ionic framework in
-                action, and a demonstration of proper code use.
-              </p>
-            </IonSlide>
+        <IonSlides ref={slideRef} onIonSlideWillChange={handleSlideChangeStart} pager={false}>
+          <IonSlide>
+            <img src="assets/img/ica-slidebox-img-1.png" className="slide-image" />
+            <h2 className="slide-title">
+              Welcome to <b>ICA</b>
+            </h2>
+            <p>
+              The <b>ionic conference app</b> is a practical preview of the ionic framework in action, and a demonstration of proper code use.
+            </p>
+          </IonSlide>
 
-            <IonSlide>
-              <img alt="what" src="/assets/img/ica-slidebox-img-2.png" className="slide-image"/>
-              <h2 className="slide-title">
-                What is Ionic?
-              </h2>
-              <p>
-                <b>Ionic Framework</b> is an open source SDK that enables developers to build
-                high quality mobile apps with web technologies like HTML, CSS, and JavaScript.
-              </p>
-            </IonSlide>
+          <IonSlide>
+            <img src="assets/img/ica-slidebox-img-2.png" className="slide-image" />
+            <h2 className="slide-title">What is Ionic?</h2>
+            <p>
+              <b>Ionic Framework</b> is an open source SDK that enables developers to build high quality mobile apps with web technologies like HTML, CSS, and JavaScript.
+            </p>
+          </IonSlide>
 
-            <IonSlide>
-              <img alt="what" src="/assets/img/ica-slidebox-img-3.png" className="slide-image"/>
-              <h2 className="slide-title">
-                What is Ionic Platform?
-              </h2>
-              <p>
-                The <b>Ionic Platform</b> is a cloud platform for managing and scaling Ionic
-                apps with integrated services like push notifications, native builds,
-                user auth, and live updating.
-              </p>
-            </IonSlide>
+          <IonSlide>
+            <img src="assets/img/ica-slidebox-img-3.png" className="slide-image" />
+            <h2 className="slide-title">What is Ionic Appflow?</h2>
+            <p>
+              <b>Ionic Appflow</b> is a powerful set of services and features built on top of Ionic Framework that brings a totally new level of app development agility to mobile dev teams.
+            </p>
+          </IonSlide>
 
-            <IonSlide>
-              <img alt="ready" src="/assets/img/ica-slidebox-img-4.png" className="slide-image"/>
-              <h2 className="slide-title">
-                Ready to Play?
-              </h2>
-              <IonButton fill="clear" onClick={this.endTutorial}>
-                Continue
-                <IonIcon slot="end" name="arrow-forward"></IonIcon>
-              </IonButton>
-            </IonSlide>
-          </IonSlides>
-        </IonContent>
-      </IonPage>
-    );
-  }
+          <IonSlide>
+            <img src="assets/img/ica-slidebox-img-4.png" className="slide-image" />
+            <h2 className="slide-title">Ready to Play?</h2>
+            <IonButton fill="clear" onClick={startApp}>
+              Continue
+              <IonIcon slot="end" icon={arrowForward} />
+            </IonButton>
+          </IonSlide>
+        </IonSlides>
+      </IonContent>
+    </IonPage>
+  );
 };
 
-const mapDispatchToProps = {
-  sawTutorial: () => actions.user.sawTutorial(),
-}
-
-export default withRouter(connect(
-  null,
-  mapDispatchToProps
-)(Tutorial));
+export default connect<OwnProps, {}, DispatchProps>({
+  mapDispatchToProps: ({
+    setHasSeenTutorial
+  }),
+  component: Tutorial
+});
